Two pointers approach | TC O(N) | Easy | Leetcode
Given a string consisting of only 'a's and 'b's, your task is to find the maximum distance between any two 'a's and the minimum distance between any two 'b's in the string. Distance should be measured by the absolute difference in indices (positions) of the characters. Then, return the sum of the maximum distance between two 'a's and the minimum distance between two 'b's.